+++ core-api/src/main/java/Hongik_SafeMap_Server/domain/disaster_report/service/DisasterReportService.java
@@ -292,7 +292,7 @@ public class DisasterReportService {
         }
 
         DisasterReportEvaluation evaluation = evaluationRepository.findDisasterReportEvaluationById(disasterReportId);
-        long accusationCount = accusationRepository.countByDisasterReportId(disasterReportId);
+        int accusationCount = accusationRepository.countByDisasterReportId(disasterReportId);
 
         return DisasterReportStatisticsResponse.of(disasterReportId, evaluation, accusationCount);
     }
